requirements.txt
PyMySQL==0.6.2
aniso8601==0.83
lxml==3.3.5
mock==1.0.1
progress==1.2
pyzmq==14.3.1
requests==2.2.1
wsgiref==0.1.2
